[DDW-809] Fix issues
