--- source/renderer/app/analytics/AnalyticsTracker.ts
@@ -1,6 +1,0 @@
-export interface AnalyticsTracker {
-  enableTracking(): Promise<void>;
-  disableTracking(): void;
-  sendPageNavigationEvent(pageTitle: string): void;
-  sendEvent(category: string, name: string, action?: string): void;
-}
