+++ source/renderer/app/analytics/MatomoAnalyticsTracker.ts
@@ -1,17 +1,16 @@
-import { AnalyticsAcceptanceStatus } from '.';
+import { AnalyticsAcceptanceStatus, AnalyticsTracker } from '.';
 import { AnalyticsClient } from './types';
 import { Environment } from '../../../common/types/environment.types';
 import LocalStorageApi from '../api/utils/localStorage';
 import { MatomoClient } from './MatomoClient';
 import { NoopAnalyticsClient } from './noopAnalyticsClient';
-import { AnalyticsTracker } from './AnalyticsTracker';
 
 export class MatomoAnalyticsTracker implements AnalyticsTracker {
   #analyticsClient: AnalyticsClient;
 
   constructor(
     private environment: Environment,
-    private localStorage: LocalStorageApi
+    private localStorageApi: LocalStorageApi
   ) {
     this.#analyticsClient = NoopAnalyticsClient;
     this.#enableTrackingIfAccepted();
@@ -20,7 +19,7 @@ export class MatomoAnalyticsTracker implements AnalyticsTracker {
   async enableTracking() {
     this.#analyticsClient = new MatomoClient(
       this.environment,
-      await localStorage.getUserID()
+      await this.localStorageApi.getUserID()
     );
   }
 
@@ -38,7 +37,7 @@ export class MatomoAnalyticsTracker implements AnalyticsTracker {
 
   async #enableTrackingIfAccepted() {
     const analyticsAccepted =
-      (await this.localStorage.getAnalyticsAcceptance()) ===
+      (await this.localStorageApi.getAnalyticsAcceptance()) ===
       AnalyticsAcceptanceStatus.ACCEPTED;
 
     if (this.environment.analyticsFeatureEnabled && analyticsAccepted) {
