+++ source/renderer/app/analytics/NoopAnalyticsTracker.ts
@@ -1,6 +1,10 @@
-export const NoopAnalyticsTracker = {
-  async enableTracking() {},
-  disableTracking() {},
-  sendPageNavigationEvent(pageTitle: string) {},
-  sendEvent(category: string, name: string) {},
-};
+import { AnalyticsTracker } from '.';
+
+class NoopAnalyticsTracker implements AnalyticsTracker {
+  async enableTracking() {}
+  disableTracking() {}
+  sendPageNavigationEvent(pageTitle: string) {}
+  sendEvent(category: string, name: string) {}
+}
+
+export const noopAnalyticsTracker = new NoopAnalyticsTracker();
