+++ source/renderer/app/analytics/types.ts
@@ -8,3 +8,10 @@ export enum AnalyticsAcceptanceStatus {
   ACCEPTED = 'ACCEPTED',
   REJECTED = 'REJECTED',
 }
+
+export interface AnalyticsTracker {
+  enableTracking(): Promise<void>;
+  disableTracking(): void;
+  sendPageNavigationEvent(pageTitle: string): void;
+  sendEvent(category: string, name: string, action?: string): void;
+}
